dom/tests/mochitest/pointerlock/file_userPref.html

<!DOCTYPE HTML>
<html>
  <!--
  https://bugzilla.mozilla.org/show_bug.cgi?id=633602
  -->
  <head>
    <title>Test for Bug 633602</title>
    <script type="text/javascript"
            src="/tests/SimpleTest/SimpleTest.js"></script>
    <script src="/tests/SimpleTest/EventUtils.js"
            type="text/javascript"></script>
    <script type="application/javascript"
            src="mouselock_util.js"></script>
    <link rel="stylesheet" type="text/css"
          href="/tests/SimpleTest/test.css" />
  </head>
  <body onload="start();">
    <a target="_blank"
       href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
      Mozilla Bug 633602</a>
    <p id="display"></p>
    <div id="content" style="display: none">
    </div>

    <canvas id="canvas" width="150" height="150"></canvas>

    <pre id="test">
      <script type="text/javascript">
        SimpleTest.waitForExplicitFinish();

        var pointer = navigator.mozPointer;
        var canvas = document.getElementById("canvas");
        SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only",
                                  false);

        function doTests () {
          if (document.mozFullScreen) {
            if (document.mozFullScreenElement === canvas) {
              // Set Preference to True to see if lock succeeds and element is
              // in fullscreen
              SpecialPowers.setBoolPref("full-screen-api.pointer-lock.enabled",
                                        true );

              pointer.lock(canvas, function() {
                is(pointer.isLocked, true, "Mouse should be locked when " +
                   "pref setting is true and element is in fullscreen.");
                pointer.unlock();

                // Set Preference to False to see if lock fails and element is
                // in fullscreen
                SpecialPowers.setBoolPref("full-screen-api.pointer-lock.enabled",
                                          false);

                pointer.lock(canvas, function() {}, function() {
                  is(pointer.isLocked, false, "Mouse shouldn't be locked " +
                     "when pref setting is false and element is in fullscreen");
                  document.removeEventListener("mozfullscreenchange", doTests,
                                                false);
                  document.mozCancelFullScreen();
                  SpecialPowers.setBoolPref("full-screen-api.pointer-lock.enabled",
                                            true);
                  SimpleTest.finish();
                });
              });
            }
          }
        }

        document.addEventListener("mozfullscreenchange", doTests, false);

        function start() {
          pointer.lock(canvas, function() {}, function() {
            is(pointer.isLocked, false, "Mouse lock can't be locked if " +
               "user preference is set to true.");
            canvas.mozRequestFullScreen();
          });
        }
      </script>
    </pre>
  </body>
</html>
